+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v1.2.0 (2019-10-15)
+* Add support for Microsoft Azure Speech-to-Text Cloud service.
+* Remove Nightly as a selectable git branch.
+
 ## v1.1.0 (2019-10-15)
 **Changes:**
 * Modifications to support different transcription services.